requirements.txt
constructs>=10.0.0,<11.0.0
botocore
boto3
pytest
black
pylint
aws-cdk-lib==2.176.0
